demo(fixed)/demo2/requirements.txt
# Core ML stack
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0

# Feature extraction
angr>=9.2.0
capstone>=5.0.0
pefile>=2023.2.7
yara-python>=4.3.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard & API
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

